Update process for user to change email
When visiting the identity/emails/edit page the form would be
shown even if the user was not verified. After fixing that, other
problems appeared, related to the callback in the User model
prematurely setting the verified status to false when submitting
the form with an incorrect password. The solution was
to use a before_update callback instead of before_validation.

There was also no system test for a user attempting to change their
email address but using the wrong password. Since that is related
to this problem I have added that as well.